+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-21T08:27:58.801Z",
+    "2026-06-22T05:54:16.909Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {